requirements.txt
absl-py==2.1.0
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1734469393482/work
aiohttp @ file:///croot/aiohttp_1734687138658/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs @ file:///croot/attrs_1734533101012/work
av==14.1.0
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1754231422783/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets @ file:///croot/datasets_1716911606380/work
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.15
dill @ file:///croot/dill_1715094664823/work
efficientnet-pytorch @ file:///home/conda/feedstock_root/build_artifacts/efficientnet-pytorch_1639403373671/work
einops==0.8.0
et_xmlfile==2.0.0
executing==2.2.0
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fonttools==4.55.4
frozenlist @ file:///croot/frozenlist_1730902802621/work
fsspec==2024.12.0
future==1.0.0
fvcore==0.1.5.post20221221
grpcio==1.70.0
huggingface-hub==0.27.1
humanize==4.11.0
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1738273805233/work
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1737102630951/work
iopath==0.1.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1730902924303/work
joblib @ file:///croot/joblib_1718217211762/work
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1736366905094/work
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml_collections==1.0.0
monai==1.5.0
moviepy==2.2.1
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
multidict @ file:///croot/multidict_1730905498140/work
multiprocess @ file:///work/perseverance-python-buildout/croot/multiprocess_1698870874259/work
munch @ file:///home/conda/feedstock_root/build_artifacts/munch_1734441240299/work
natsort @ file:///tmp/build/80754af9/natsort_1611692802956/work
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1737039604450/work
nibabel==5.3.2
numexpr @ file:///croot/numexpr_1730215937391/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
omegaconf==2.3.0
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp312-cp312-linux_x86_64.whl#sha256=57b66702d418720ec8483f7c4ec7c08d41815316ad7ce09d5b7bbc34eefcfdfd
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow @ file:///croot/pillow_1734430590456/work
platformdirs==4.3.6
portalocker==3.1.1
pretrainedmodels @ file:///home/conda/feedstock_root/build_artifacts/pretrainedmodels_1619808183723/work
proglog @ file:///home/conda/feedstock_root/build_artifacts/proglog_1728452852274/work
prompt_toolkit==3.0.50
propcache @ file:///croot/propcache_1732303986938/work
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow @ file:///croot/pyarrow_1737973085896/work/python
pycpd==2.0.0
pydicom==3.0.1
Pygments==2.19.1
pynrrd==1.1.1
pyparsing==3.2.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.1.1
pytorchvideo==0.1.5
pytz @ file:///croot/pytz_1713974312559/work
PyWavelets==1.8.0
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.2.1
regex @ file:///croot/regex_1736540786412/work
requests @ file:///croot/requests_1730999120400/work
rich==13.9.4
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn @ file:///croot/scikit-learn_1737988764076/work
scipy @ file:///croot/scipy_1737122901252/work/dist/scipy-1.15.1-cp312-cp312-linux_x86_64.whl#sha256=7b1b0001c943e119509d010e5cbb4950dd6e325a83ca2510a25f139dac6d66a8
seaborn==0.13.2
segment-anything==1.0
segmentation_models_pytorch @ file:///home/conda/feedstock_root/build_artifacts/segmentation-models-pytorch_1744915137288/work
setuptools==72.1.0
shapely==2.0.6
shellingham==1.5.4
SimpleITK==2.4.1
six==1.17.0
slicerio==1.1.1
spyder-kernels==2.5.2
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tifffile==2025.1.10
timm==1.0.14
tokenizers @ file:///croot/tokenizers_1732129617059/work
torch==2.5.1
torchaudio==2.5.1
torchio==0.20.3
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers @ file:///croot/transformers_1729767158381/work
triton==3.1.0
typer==0.15.1
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1737133630106/work
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.2
wurlitzer==3.1.1
xxhash @ file:///croot/python-xxhash_1737039434400/work
yacs==0.1.8
yarl @ file:///croot/yarl_1732546845924/work
